Move EventsButton to Atoms
