+++ src/NavBar/NavBar.tsx
@@ -15,7 +15,7 @@ import {
 
 import { Menu as MenuIcon, Language as LanguageIcon } from '@material-ui/icons';
 import { withStyles } from '@material-ui/core/styles';
-import EventsButton from './EventsButton';
+import EventsButton from '../components/Atoms/EventsButton/EventsButton';
 
 import i18n from '../i18n';
 import ResponsiveImage from '../ResponsiveImage/ResponsiveImage';
